Fix some type mismatches when using 64-bit size_t and uintptr_t.
These only produce warnings, not errors.

+++ Sources/Plasma/FeatureLib/pfConsole/pfDispatchLog.cpp
@@ -92,7 +92,7 @@ void plDispatchLog::LogStatusBarChange(const char* name, const char* action)
     memset(&mbi, 0, sizeof(MEMORY_BASIC_INFORMATION));
 
     // Note: this will return shared mem too on Win9x.  There's a way to catch that, but it's too slow -Colin
-    uint32_t processMemUsed = 0;
+    size_t processMemUsed = 0;
     void* curAddress = nullptr;
     while (VirtualQuery(curAddress, &mbi, sizeof(MEMORY_BASIC_INFORMATION)) == sizeof(MEMORY_BASIC_INFORMATION))
     {
